refactor: 인수테스트에 중복되어는 setup 코드 baseAcceptance 로 이동
