+++ src/test/java/subway/LineAcceptanceTest.java
@@ -1,17 +1,11 @@
 package subway;
 
 import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import io.restassured.specification.RequestSpecification;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.annotation.DirtiesContext;
 import subway.web.request.LineCreateRequest;
 import subway.web.request.LineUpdateRequest;
 import subway.web.response.StationResponse;
@@ -21,21 +15,9 @@ import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.test.annotation.DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD;
 
 @DisplayName("노선 관련 기능")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = BEFORE_EACH_TEST_METHOD)
-class LineAcceptanceTest {
-
-    private static RequestSpecification REQUEST_SPEC;
-
-    @BeforeEach
-    public void setUp() {
-        RequestSpecBuilder reqBuilder = new RequestSpecBuilder();
-        reqBuilder.setContentType(ContentType.JSON);
-        REQUEST_SPEC = reqBuilder.build();
-    }
+class LineAcceptanceTest extends BaseAcceptance {
 
     /**
      * Given 지하철역을 생성하고
